refactor: removed unnecessary sequence package under the masks and moved the padding and lookahead masks to their respective files under the main masks package.
This helps the api cleaner and simpler
All tests are passing
