+++ transformerx/layers/masks/__init__.py
@@ -1,2 +1,3 @@
-from transformerx.layers.masks.sequence_masks.generic import LookAheadMask, PaddingMask
 from transformerx.layers.masks.core import BaseMask
+from transformerx.layers.masks.padding import PaddingMask
+from transformerx.layers.masks.lookahead import LookAheadMask
